+++ src/utils/file_upload.rs
@@ -412,7 +412,7 @@ fn upload_files_chunked(
     pb.set_style(progress_style);
 
     let view = ByteView::open(archive.path())?;
-    let (checksum, checksums) = get_sha1_checksums(&view, options.chunk_size)?;
+    let (checksum, checksums) = get_sha1_checksums(&view, options.chunk_size as usize)?;
     let mut chunks = view
         .chunks(options.chunk_size as usize)
         .zip(checksums.iter())
